ml/requirements-ml.txt
torch>=2.2.0
torchvision>=0.17.0
pandas>=2.2.0
scikit-learn>=1.4.0
pillow>=10.0.0
tqdm>=4.66.0
